% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_polygon.R
\name{assess_polygon}
\alias{assess_polygon}
\title{Assess the geometry of a polygon}
\usage{
assess_polygon(polygon_path, survey_points, settlement_points, ornl)
}
\arguments{
\item{polygon_path}{the relative file path to the polygon shapefile}

\item{survey_points}{Optional. The survey points where the delineation should occur}

\item{settlement_points}{Optional. The settlement type points layer}

\item{ornl}{Optional. A shapefile of the previously surveyed locations}
}
\description{
Runs a number of diagnostics on the delineated polygons to check the geometry.
}
