% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_polygon.R
\name{assess_polygons}
\alias{assess_polygons}
\title{Assess Multiple Polygons}
\usage{
assess_polygons(polygons, survey_points, settlement_points, ornl,
  merged_polygons = NULL, progress = FALSE)
}
\arguments{
\item{survey_points}{Optional. The survey points where the delineation should occur}

\item{settlement_points}{Optional. The settlement type points layer}

\item{ornl}{Optional. A shapefile of the previously surveyed locations}

\item{merged_polygons}{Optional. A SpatialPolygonDataFrame of the delineated polygons created from the \code{merge_delineated_polygons} function.}

\item{progress}{should messages be displayed for the progress through the polygons? Default is TRUE}
}
\description{
A wrapper for the \code{assess_polygon} function to run the function for multiple polygons
}
