% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_crs.R
\name{check_crs}
\alias{check_crs}
\title{Validate the CRS of a shapefile}
\usage{
check_crs(input,
  crs_ref = "+proj=longlat +datum=WGS84 +no_defs +ellps=WGS84 +towgs84=0,0,0")
}
\description{
Validate the CRS of a shapefile
}
