% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_reports.R
\name{create_reports}
\alias{create_reports}
\title{Build a diagnostics summary report}
\usage{
create_reports(polygons = polygonName, out_dir = "Reports",
  input_file = NULL)
}
\arguments{
\item{polygons}{the name of the filepath to the polygon}

\item{out_dir}{the output directory}
}
\description{
Will create a summary report for each summary report.
}
\author{
Michael Harper
}
