% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_id.R
\name{extract_id}
\alias{extract_id}
\title{Extract ID from polygon name}
\usage{
extract_id(filepath)
}
\arguments{
\item{filepath}{a filepath or file name of a shapefile}
}
\description{
For a specified polygon file, extract the three digit ID. It is expected that
 the polygon file has the name format "mh_030_polygon.shp".
}
