% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_delineated_polygons.R
\name{merge_delineated_polygons}
\alias{merge_delineated_polygons}
\title{Merges a list of delineated polygons into a single spatial object}
\usage{
merge_delineated_polygons(polygons)
}
\arguments{
\item{polygons}{a list of polygons}
}
\value{
A SpatialPolygonsDataFrame with the identifier column ID
}
\description{
Merges a list of delineated polygons into a single spatial object
}
