% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_polygon.R
\name{assess_polygon}
\alias{assess_polygon}
\title{Assess a single polygon}
\usage{
assess_polygon(polygon_path, survey_points, settlement_points, ornl,
  merged_polygons = NULL, progress = TRUE)
}
\arguments{
\item{polygon_path}{the relative file path to the polygon shapefile}

\item{survey_points}{Optional. The survey points where the delineation should occur}

\item{settlement_points}{Optional. The settlement type points layer}

\item{ornl}{Optional. A shapefile of the previously surveyed locations}

\item{merged_polygons}{Optional. A SpatialPolygonDataFrame of the delineated polygons created from the \code{merge_delineated_polygons} function.}

\item{progress}{should messages be displayed for the progress through the polygons? Default is TRUE}
}
\description{
Internal function that is used to diagnose a single Polygon. Please use the
 \code{\link{assess_polygons}} function which is designed to use this function
 across a list of polygons.
}
