% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_polygons.R
\name{list_polygons}
\alias{list_polygons}
\title{List Polygons within a directory}
\usage{
list_polygons(..., recursive = TRUE)
}
\arguments{
\item{...}{a list of values to be concatenated into a filepath}
}
\value{
a list of polygons
}
\description{
Returns a list of filepaths to shapefiles. This will recursively search any
 subdirectories
}
\examples{
list_polygons(system.file("extdata", package = "polyCheck"), "polygons")

}
\author{
Michael Harper
}
